requirements.txt
pandas
google-genai
pytest
autopep8
flake8